fix dialyzer with temporary workaround
This commit temporarily fixes the problem raised by dialyzer as "Function NotificationController.send/2 has no local return”.

The issue is that conn parameter is of Plug.Conn type, which means a struct with :body_params field defined as a map.
OpenApiSpex library changes this map to user-defined struct (complying with user-provided schema definition).
Since, from Elixir types point of view, this struct and map are actually of different types, when we pass the conn object
to any of Plug functions, which are specified to take arguments of Plug.Conn type, we're technically not passing Plug.Conn type,
because the type of :body_params is wrong. Dialyzer assumes then that call will fail. Alternatively, corresponding story
applies to RequestDecoder.decode/1 function, however, simultaneously only one of this cases is being reported by the tool.

The workaround (consisting of :body_params type change from struct to map) will be immediately removed as the CastAndValidate plug
will be reimplemented, as intended, to use :private key to store incoming schema instead of :body_params one.