Unify telemetry handlers
There's no real difference between events from APNS and FCM, two
identical modules are redundant for this. To telemetry, the service is
already part of the metadata, it can be simply extracted from there and
passed to exometer as exometer itself needs, but the event can be kept
generic. This will hopefully help with the future event logic regarding
telemetry_metrics.