MIM-868 Handle APIv1 requests (phoenix controller) (#150)
* align & implement MoongoosePush.Notification behaviour

* translate APIv1 requests to MongoosePush format; add unit tests

* add more functional tests of APIv1.NotificationController

* correct spec

* add more specs in RequestDecoder

* update forgotten function name

* mix format

* mix format

* fix dialyzer with temporary workaround

This commit temporarily fixes the problem raised by dialyzer as "Function NotificationController.send/2 has no local return”.

The issue is that conn parameter is of Plug.Conn type, which means a struct with :body_params field defined as a map.
OpenApiSpex library changes this map to user-defined struct (complying with user-provided schema definition).
Since, from Elixir types point of view, this struct and map are actually of different types, when we pass the conn object
to any of Plug functions, which are specified to take arguments of Plug.Conn type, we're technically not passing Plug.Conn type,
because the type of :body_params is wrong. Dialyzer assumes then that call will fail. Alternatively, corresponding story
applies to RequestDecoder.decode/1 function, however, simultaneously only one of this cases is being reported by the tool.

The workaround (consisting of :body_params type change from struct to map) will be immediately removed as the CastAndValidate plug
will be reimplemented, as intended, to use :private key to store incoming schema instead of :body_params one.

* tune up schemas names

* get rid of Request.SendNotification.Deep schema

* move alert and data fields to separate schemas

* introduce RequestDecoder protocol

* remove old RequestDecoder; minor changes to filenames and location

* better parsing :more field's value