Telemetry metrics (#151)
* Be more explicit about the fact that we are still exometer based

* Turn away from macros, prefer regular functions

* Clean-up module scope rules

For example, no need to prefix local functions with the module name,
and the function `name` can be private.

* No need to expand macros anymore

The module `Metrics` does not export any, so we can omit the `use`
directive for this module.

* Make `Metrics` a behaviour.

`MongoosePush.Metrics` is a behaviour now, with the corresponding
`update_` callbacks required, while `MongoosePush.Metrics, Exometer`
implements this behaviour. This will help define
`MongoosePush.Metrics.TelemetryMetrics` as an implementation for such
behaviour as well.

* Add telemetry_metrics to mix

* Unify telemetry handlers

There's no real difference between events from APNS and FCM, two
identical modules are redundant for this. To telemetry, the service is
already part of the metadata, it can be simply extracted from there and
passed to exometer as exometer itself needs, but the event can be kept
generic. This will hopefully help with the future event logic regarding
telemetry_metrics.

* Move metric files around

* Keep reducing files

* Template for telemetry_metrics event handlers

* Telemetry support
* Changed telemetry events to support multi-dimensional metrics
* Added metrics' definitions
* Added Prometheus reporter
* Added telemetry_metrics and telemetry_metrics_prometheus_core dependencies
* Added tests for the metrics

* Move telemetry_metrics defs to its own module

* Make notification events a distribution

* Change name of the push event

`[:mongoose_push, :notification, :send]` seems more telling.

* PR review changes

* Changed config files to include `:service` -> module mapping
* Changed the loading of said module from hardcoded variable to
extracted from config
* Renamed metric tags to be more informative for external users
* Fixed `MongoosePush.Service.choose_pool/2` callback (it had arity of 1
and did not take tags as an argument)
* Added file defining mock with Mox (Needed because of compile-time
restraints)
* Replaced Mock with Mox

* Ignore dialyzer error due to wrong library specs

TelemetryMetricsPrometheus.Core exports incomplete metrics for the
child_spec options, which makes our call fail. Tell dialyzer to ignore
this.

* Mox doesn't need compile-time dependencies

* Prepare tag_values earlier

* Fix exometer metadata naming expectations

* Apply PR review comments

* Configuration logic handled in the `Application` module

* Apply review

* Refactor test suite

* Use second as timeunit as per Prometheus recommendation

* Fix TelemetryMetricsPrometheus startup race condition

* Update test/unit/mongoose_push_telemetry_metrics_test.exs

Co-authored-by: Nelson Vides <nelson.vides@erlang-solutions.com>

Co-authored-by: Nelson Vides <nelson.vides@erlang-solutions.com>
Co-authored-by: Rafał Słota <rafal.slota@erlang-solutions.com>